LayoutTests/css3/viewport-percentage-lengths/viewport-percentage-lengths-resize-expected.txt

Test that viewport lengths and font sizes properly resize with the viewport.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
PASS innerWidth is 800
PASS innerHeight is 600
PASS getComputedStyle(test).width is "400px"
PASS getComputedStyle(testfontsize).fontSize is "30px"
PASS getComputedStyle(testcalc).width is "800px"
PASS getComputedStyle(testfontsizecalc).fontSize is "60px"
PASS getComputedStyle(testpseudo, ':after').marginLeft is "120px"
PASS getComputedStyle(testpseudo, ':after').paddingRight is "200px"
PASS getComputedStyle(testpseudocalc, ':after').marginLeft is "240px"
PASS getComputedStyle(testpseudocalc, ':after').paddingRight is "400px"
PASS innerWidth is 900
PASS innerHeight is 600
PASS getComputedStyle(test).width is "450px"
PASS getComputedStyle(testfontsize).fontSize is "30px"
PASS getComputedStyle(testcalc).width is "900px"
PASS getComputedStyle(testfontsizecalc).fontSize is "60px"
PASS getComputedStyle(testpseudo, ':after').marginLeft is "120px"
PASS getComputedStyle(testpseudo, ':after').paddingRight is "225px"
PASS getComputedStyle(testpseudocalc, ':after').marginLeft is "240px"
PASS getComputedStyle(testpseudocalc, ':after').paddingRight is "450px"
PASS innerWidth is 900
PASS innerHeight is 640
PASS getComputedStyle(test).width is "450px"
PASS getComputedStyle(testfontsize).fontSize is "32px"
PASS getComputedStyle(testcalc).width is "900px"
PASS getComputedStyle(testfontsizecalc).fontSize is "64px"
PASS getComputedStyle(testpseudo, ':after').marginLeft is "128px"
PASS getComputedStyle(testpseudo, ':after').paddingRight is "225px"
PASS getComputedStyle(testpseudocalc, ':after').marginLeft is "256px"
PASS getComputedStyle(testpseudocalc, ':after').paddingRight is "450px"
PASS innerWidth is 500
PASS innerHeight is 640
PASS getComputedStyle(test).width is "250px"
PASS getComputedStyle(testfontsize).fontSize is "32px"
PASS getComputedStyle(testcalc).width is "500px"
PASS getComputedStyle(testfontsizecalc).fontSize is "64px"
PASS getComputedStyle(testpseudo, ':after').marginLeft is "100px"
PASS getComputedStyle(testpseudo, ':after').paddingRight is "160px"
PASS getComputedStyle(testpseudocalc, ':after').marginLeft is "200px"
PASS getComputedStyle(testpseudocalc, ':after').paddingRight is "320px"
PASS innerWidth is 800
PASS innerHeight is 600
PASS getComputedStyle(test).width is "400px"
PASS getComputedStyle(testfontsize).fontSize is "30px"
PASS getComputedStyle(testcalc).width is "800px"
PASS getComputedStyle(testfontsizecalc).fontSize is "60px"
PASS getComputedStyle(testpseudo, ':after').marginLeft is "120px"
PASS getComputedStyle(testpseudo, ':after').paddingRight is "200px"
PASS getComputedStyle(testpseudocalc, ':after').marginLeft is "240px"
PASS getComputedStyle(testpseudocalc, ':after').paddingRight is "400px"
